// videoeditor/src/components/timeline/types.ts
export interface BaseScrubber {
  id: string;
  mediaType: "video" | "image" | "audio" | "text" | "groupped_scrubber" | "subtitle";
  mediaUrlLocal: string | null; // null for text and subtitle
  mediaUrlRemote: string | null; // for subtitle, this can hold the path like /api/subtitles/foo.json
  media_width: number; // width of the media in pixels
  media_height: number; // height of the media in pixels

  text: TextProperties | null;
  subtitleData: any | null; // Parsed JSON data for subtitle
  groupped_scrubbers: ScrubberState[] | null; // null for not grouped
  //  groupped_scrubber_transitions: Transition[] | null; // null for no transitions / not groupped scrubbers [this is written to help with deepcopy]

  // transitions are managed using the right transition id, as in what to add to the right. Convenient to think of. Left one is for the initial transition, first scrubber intro. we won't use it anywhere else.
  // for a middle transition, you will only see its information in the left scrubber.
  left_transition_id: string | null; // only use this for the first scrubber intro
  right_transition_id: string | null; // this is what you use everywhere
}
// ...
export interface TextProperties {
  textContent: string; // Only present when mediaType is "text"
  fontSize: number;
  fontFamily: string;
  color: string;
  textAlign: "left" | "center" | "right";
  fontWeight: "normal" | "bold";
  template: "normal" | "glassy" | null;          // template uses tiktok style pages. null for normal text. templates might override the text properties.
}

// state of the scrubber in the media bin
export interface MediaBinItem extends BaseScrubber {
  name: string;
  durationInSeconds: number; // For media, to calculate initial width

  // Upload tracking properties
  uploadProgress: number | null; // 0-100, null when upload complete
  isUploading: boolean; // True while upload is in progress
}

// state of the scrubber in the timeline
export interface ScrubberState extends MediaBinItem {
  left: number; // in pixels (for the scrubber in the timeline)
  y: number; // track position (0-based index)
  width: number; // width is a css property for the scrubber width
  sourceMediaBinId: string; // ID of the media bin item this scrubber was created from

  // the following are the properties of the scrubber in <Player>
  left_player: number;
  top_player: number;
  width_player: number;
  height_player: number;
  is_dragging: boolean;

  // for video scrubbers (and audio in the future)
  trimBefore: number | null; // in frames
  trimAfter: number | null; // frames trimmed from the end of the source media
}